% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_sensitivity.R
\name{generate_pricing_stats}
\alias{generate_pricing_stats}
\title{Generate a dataframe of pricing statistics
over a sliding window from the minimum amount to
the maximum.}
\usage{
generate_pricing_stats(
  surveydata,
  too_expensive,
  bargain,
  getting_expensive,
  too_cheap
)
}
\description{
Generate a dataframe of pricing statistics
over a sliding window from the minimum amount to
the maximum.
}
